Fix and re-enable QSFragmentTest
QSFragmentTest has been disabled for a long time, as it is failing.
QSFragment is a crucial component of SysUI and should have tests.

The fix was to make sure all dependencies are correctly set up.
This also applied to the layout inflater and views. By default the
code was inflating the real view, and some of the sub-views have
dependencies that were not there.

Fixes: 227167398
Test: QSFragmentTest passes
Change-Id: I0a0a23c4e5b023d1178d9f96a561730051cd98aa