Clear the preferred top focusable root task when moving
... a focusable and inivisble root task to top.

A focusable root task was moved to top, but it was not the
top focusable root task because the task was force-hidden
and not visible. The mPreferredTopFocusableRootTask should
be clear in that case, so that the top focusable root task
can be recalculated to the topmost root task once it became
visible.

Bug: 228092608
Test: repro on the car emulator
Change-Id: I5e645b5148ab03014a95ac92d540e3c6f242bbfe